+++ public/data/boards.json
@@ -38,6 +38,19 @@
       "file_path": null,
       "profile_image_path": "/images/default.png",
       "post_id": 37
+    },
+    {
+      "post_title": "제목",
+      "post_content": "히히히히",
+      "user_id": 1,
+      "nickname": "테스트",
+      "created_at": "2024. 5. 7. 오후 10:50:53",
+      "updated_at": "2024. 5. 7. 오후 10:50:53",
+      "comment_count": "0",
+      "hits": "1",
+      "file_path": "/uploads/back.png",
+      "profile_image_path": "/images/default.png",
+      "post_id": 38
     }
   ]
 }
