+++ public/data/comments.json
@@ -9,6 +9,36 @@
       "updated_at": "2024. 5. 7. 오후 10:35:42",
       "profile_image_path": "/images/default.png",
       "comment_id": 38
+    },
+    {
+      "comment_content": "asfsf\n",
+      "post_id": 36,
+      "user_id": 1,
+      "nickname": "테스트",
+      "created_at": "2024. 5. 7. 오후 10:41:53",
+      "updated_at": "2024. 5. 7. 오후 10:41:53",
+      "profile_image_path": "/images/default.png",
+      "comment_id": 39
+    },
+    {
+      "comment_content": "dfsff",
+      "post_id": 36,
+      "user_id": 1,
+      "nickname": "테스트",
+      "created_at": "2024. 5. 7. 오후 10:41:54",
+      "updated_at": "2024. 5. 7. 오후 10:41:54",
+      "profile_image_path": "/images/default.png",
+      "comment_id": 40
+    },
+    {
+      "comment_content": "aff",
+      "post_id": 36,
+      "user_id": 1,
+      "nickname": "테스트",
+      "created_at": "2024. 5. 7. 오후 10:42:10",
+      "updated_at": "2024. 5. 7. 오후 10:42:10",
+      "profile_image_path": "/images/default.png",
+      "comment_id": 41
     }
   ]
 }
